backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-multipart==0.0.20
pillow==11.1.0
numpy==2.2.1
pandas==2.2.3
opencv-python==4.10.0.84
scikit-image==0.24.0
torch==2.5.1
torchvision==0.20.1
timm==1.0.12
monai==1.4.0
scikit-learn==1.6.0
tqdm==4.67.1
aiofiles==24.1.0
